requirements.txt
python-telegram-bot[ext]==21.10
google-genai==1.12.1
aiosqlite==0.21.0
openpyxl==3.1.5
aiosmtplib==3.0.2
apscheduler==3.11.0
python-dotenv==1.1.0
aiohttp==3.11.14
